migrate content of entities tab to separate component
