+++ packages/client/src/pages/Stats/EntitiesTab/StatsChart/StatsChart.tsx
@@ -16,13 +16,13 @@ import {
 
 import api from "api";
 import { useTheme } from "styled-components";
-import { OTHERS_KEY } from "../constants";
 import {
   ChartDataPoint,
   getCategoryMap,
   getDataCategories,
+  OTHERS_KEY,
   transformDataForChart,
-} from "../utils";
+} from "../../utils";
 import {
   StyledChartWrapper,
   StyledCustomTooltip,
