+++ packages/client/src/pages/Stats/EntitiesTab/StatsChart/StatsChartStyles.tsx
