+++ packages/client/src/pages/Stats/EntitiesTab/StatsTable/StatsTable.tsx
@@ -3,8 +3,8 @@ import { useQuery } from "@tanstack/react-query";
 import api from "api";
 import { useMemo } from "react";
 import { Column, useTable } from "react-table";
-import { TABLE_PADDING } from "../constants";
-import { getDataCategories, transformDataForTable } from "../utils";
+import { TABLE_PADDING } from "../../constants";
+import { getDataCategories, transformDataForTable } from "../../utils";
 import {
   StyledTable,
   StyledTableContainer,
