+++ packages/client/src/pages/Stats/EntitiesTab/StatsTable/StatsTableStyles.tsx
